+++ src/test/scala/generators/AstGenerators.scala
@@ -28,7 +28,7 @@ trait AstGenerators extends CommonGenerators {
     * @see
     *   [[cl.ravenhill.scum.generators.CommonGenerators.generateLong]]
     */
-  def generateNumericLiteral(value: Gen[Int] = generateLong()): Gen[NumericLiteral[String]] = for {
+  def generateNumericLiteral(value: Gen[Long] = generateLong()): Gen[NumericLiteral[String]] = for {
     v <- value
     n <- Gen.const(NumericLiteral[String](v))
   } yield n
